Last review and fixed those errors
even though you let the user know that the * means the field is required, it is better if they get an error message describing the why, the error for example in the credit card information, or zip code is also a requirement for exceeds expectations which i definitely believe you are going for ! :100:  here's the link if you need a refresher https://teamtreehouse.com/projects/interactive-form#how-you-ll-be-graded
Make sure the to hide the color dropdown menu when there's no design selected.
The email field allow me to write something like this some@email.!@#$  a quick adjust to your regex should fix that.